Fix inconsistent token source
