+++ packages/dara-core/js/auth/index.tsx
@@ -1,13 +1,5 @@
 export { default as AuthCtx, useAuthCtx, useSessionToken } from './auth-context';
-export {
-    verifySessionToken,
-    useUser,
-    useSession,
-    handleAuthErrors,
-    DARA_JWT_TOKEN,
-    getSessionToken,
-    revokeSession,
-} from './auth';
+export { verifySessionToken, useUser, useSession, handleAuthErrors, getSessionToken, revokeSession } from './auth';
 export { default as AuthWrapper } from './auth-wrapper';
 export { default as BasicAuthLogin } from './basic/basic-auth-login';
 export { default as BasicAuthLogout } from './basic/basic-auth-logout';
