+++ packages/dara-core/js/shared/utils/embed.tsx
@@ -1,3 +1,5 @@
+export const DARA_JWT_TOKEN = 'dara-jwt-token';
+
 /**
  * Whether the Dara app is embedded within an IFrame
  */
@@ -19,4 +21,27 @@ export function getEmbedTokenKey(): string {
  */
 export function getEmbedToken(): string {
     return localStorage.getItem(getEmbedTokenKey());
+}
+
+/**
+ * Get the jwt token. Uses the embedded token if the app is embedded.
+ */
+export function getToken(): string {
+    if (isEmbedded()) {
+        const embedToken = getEmbedToken();
+        return embedToken;
+    }
+
+    return localStorage.getItem(DARA_JWT_TOKEN);
+}
+
+/**
+ * Get the token key used to persist the jwt token. Uses the embedded token key if the app is embedded.
+ */
+export function getTokenKey(): string {
+    if (isEmbedded()) {
+        return getEmbedTokenKey();
+    }
+
+    return DARA_JWT_TOKEN;
 }
